+++ docs/source/release_notes/release-notes.rst
@@ -47,6 +47,12 @@ develop
            linear in terms of stream length, as it was previously.
            (`Pull Request <https://github.com/palantir/atlasdb/pull/1335>`__)
 
+    *    - |new|
+         - AtlasDB now supports stream store compression.
+           Streams can be compressed client-side by adding the ``compressStreamInClient`` option to the stream 
+           definition. Reads from the stream store will transparently decompress the data.
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/1357>`__)
+
 .. <<<<------------------------------------------------------------------------------------------------------------->>>>
 
 =======